requirements.txt
openai
python-dotenv
langchain
langchain-openai
langchain-ollama
langchain-community
langchain_google_genai
langchainhub
faiss-cpu
pypdf
pandas